requirements.txt
dj-database-url==0.4.0
django==1.8
psycopg2==2.6.1
whitenoise==2.0.6
django-registration==2.1.2
django-bootstrap3
django-bootstrap3-datetimepicker
uwsgi
python-ldap
